renamed transaction-history/index
